fix(security): sanitize le contenu social (XSS stocké) + corrige le rendu profil
FAILLE : les status_posts (réseau social) étaient insérés sans sanitization
(social.ts) et rendus en {@html} dans le feed → XSS stocké (un membre pouvait
exécuter du JS chez tous ceux qui voyaient son post). La migration 053 avait
autorisé le HTML sans ajouter de filtre.

- Extrait le sanitize() du forum (allowlist éprouvée) dans utils/sanitize.ts,
  réutilisé par forums.ts ET social.ts (un seul allowlist, plus de drift).
- social.ts : sanitize le contenu à la création (POST /status). Vérifié :
  <script>/<img onerror>/javascript: neutralisés, embeds YouTube préservés.
- Backfill prod : re-sanitize des 14 status_posts existants (6 normalisés, 0
  contenait un motif XSS actif).
- Profil : le rendu passait par {post.content} (HTML brut affiché en texte) →
  {@html} + classe nodyx-prose, comme le feed. Bug d'affichage corrigé.

Faille trouvée en corrigeant le bug d'affichage signalé par Jonathan.
